+++ src/app/account/page.tsx
@@ -1,3 +1,4 @@
+'use client';
 import * as React from 'react';
 
 import Seo from '@/components/Seo';
