+++ src/app/index/page.tsx
@@ -1,3 +1,4 @@
+'use client';
 import * as React from 'react';
 
 import Carousel, { CarouselProps } from '@/components/carousel/Carousel';
