+++ src/app/layout.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+
+import '../styles/globals.css';
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang='en'>
+      <head>
+        <link
+          rel='preload'
+          href='/fonts/inter-var-latin.woff2'
+          as='font'
+          type='font/woff2'
+          crossOrigin='anonymous'
+        />
+      </head>
+      <body>{children}</body>
+    </html>
+  );
+}
+
+/*return (
+    <Html lang='en'>
+      <Head>
+        <link
+          rel='preload'
+          href='/fonts/inter-var-latin.woff2'
+          as='font'
+          type='font/woff2'
+          crossOrigin='anonymous'
+        />
+      </Head>
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+}
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
+}
+
+export default MyApp;
+*/
