+++ src/app/logout/page.tsx
@@ -1,4 +1,6 @@
-import { useRouter } from 'next/router';
+'use client';
+
+import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 
 export default function LogoutPage() {
