+++ src/components/Seo.tsx
@@ -1,5 +1,5 @@
 import Head from 'next/head';
-import { useRouter } from 'next/router';
+import { usePathname } from 'next/navigation';
 
 const defaultMeta = {
   title: 'Nex | Music Player',
@@ -22,7 +22,6 @@ type SeoProps = {
 } & Partial<typeof defaultMeta>;
 
 export default function Seo(props: SeoProps) {
-  const router = useRouter();
   const meta = {
     ...defaultMeta,
     ...props,
@@ -46,8 +45,8 @@ export default function Seo(props: SeoProps) {
       <title>{meta.title}</title>
       <meta name='robots' content={meta.robots} />
       <meta content={meta.description} name='description' />
-      <meta property='og:url' content={`${meta.url}${router.asPath}`} />
-      <link rel='canonical' href={`${meta.url}${router.asPath}`} />
+      <meta property='og:url' content={`${meta.url}${usePathname}`} />
+      <link rel='canonical' href={`${meta.url}${usePathname}`} />
       {/* Open Graph */}
       <meta property='og:type' content={meta.type} />
       <meta property='og:site_name' content={meta.siteName} />
