--- src/pages/_app.tsx
@@ -1,15 +1,0 @@
-import { AppProps } from 'next/app';
-
-import '@/styles/globals.css';
-
-import Layout from '@/components/layout/Layout';
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
-  );
-}
-
-export default MyApp;
