+++ tailwind.config.ts
@@ -2,7 +2,7 @@ import type { Config } from 'tailwindcss';
 import defaultTheme from 'tailwindcss/defaultTheme';
 
 export default {
-  content: ['./src/**/*.{js,jsx,ts,tsx}'],
+  content: ['./src/app/**/*.{js,jsx,ts,tsx}'],
   theme: {
     fontFamily: {
       quicksand: ['Quicksand', 'sans-serif'],
